move `update_relay_configuration` to account/relay.rs
